extract activities by WP into own endpoint
