+++ lib/api/v3/work_packages/work_packages_api.rb
@@ -26,7 +26,6 @@
 # See doc/COPYRIGHT.rdoc for more details.
 #++
 
-require 'api/v3/activities/activity_representer'
 require 'api/v3/work_packages/work_package_representer'
 
 module API
@@ -79,36 +78,9 @@ module API
               end
             end
 
-            resource :activities do
-              helpers do
-                def save_work_package(work_package)
-                  if work_package.save
-                    journals = ::Journal::AggregatedJournal.aggregated_journals(
-                      journable: work_package)
-                    Activities::ActivityRepresenter.new(journals.last, current_user: current_user)
-                  else
-                    fail ::API::Errors::ErrorBase.create_and_merge_errors(work_package.errors)
-                  end
-                end
-              end
-
-              params do
-                requires :comment, type: String
-              end
-              post do
-                authorize({ controller: :journals, action: :new },
-                          context: @work_package.project) do
-                  raise ::API::Errors::NotFound.new
-                end
-
-                @work_package.journal_notes = params[:comment]
-
-                save_work_package(@work_package)
-              end
-            end
-
             mount ::API::V3::WorkPackages::WatchersAPI
             mount ::API::V3::Relations::RelationsAPI
+            mount ::API::V3::Activities::ActivitiesByWorkPackageAPI
             mount ::API::V3::Attachments::AttachmentsByWorkPackageAPI
             mount ::API::V3::WorkPackages::UpdateFormAPI
           end
